dashboard/backend/market_intelligence.py
import yfinance as yf
import pandas as pd
import numpy as np
from alpha_vantage.timeseries import TimeSeries
from prophet import Prophet
import ccxt
import requests
from bs4 import BeautifulSoup
from pytrends.request_builder import TrendReq
from datetime import datetime, timedelta
import asyncio
import aiohttp
from newsapi import NewsApiClient
from sklearn.ensemble import RandomForestRegressor
import logging

class MarketIntelligenceSystem:
    def __init__(self, api_keys):
        self.api_keys = api_keys
        self.setup_apis()
        
    def setup_apis(self):
        self.alpha_vantage = TimeSeries(key=self.api_keys['ALPHA_VANTAGE'])
        self.news_api = NewsApiClient(api_key=self.api_keys['NEWS_API'])
        self.pytrends = TrendReq(hl='en-US', tz=360)
        
    async def analyze_competition(self, company_name):
        """Comprehensive competitor analysis"""
        competitors = await self.identify_competitors(company_name)
        analysis = {
            'market_share': await self.analyze_market_share(competitors),
            'strategies': await self.analyze_business_strategies(competitors),
            'strengths_weaknesses': await self.analyze_swot(competitors),
            'pricing': await self.analyze_pricing_strategies(competitors),
            'technology': await self.analyze_tech_stack(competitors),
            'social_media': await self.analyze_social_presence(competitors)
        }
        return analysis

    async def predict_market_trends(self, sector, timeframe='1y'):
        """Advanced market trend prediction"""
        data = {
            'economic_indicators': await self.get_economic_indicators(),
            'sector_performance': await self.analyze_sector_performance(sector),
            'sentiment_analysis': await self.analyze_market_sentiment(),
            'technical_analysis': await self.perform_technical_analysis(),
            'ai_predictions': await self.generate_ai_forecasts()
        }
        return data

    async def crypto_market_analysis(self):
        """Cryptocurrency market analysis"""
        exchanges = ['binance', 'coinbase', 'kraken']
        analysis = {
            'market_overview': await self.analyze_crypto_market(),
            'top_opportunities': await self.identify_crypto_opportunities(),
            'risk_assessment': await self.assess_crypto_risks(),
            'volume_analysis': await self.analyze_trading_volumes(),
            'correlation_matrix': await self.generate_correlation_matrix()
        }
        return analysis

    async def stock_market_analysis(self, symbols):
        """Advanced stock market analysis"""
        analysis = {
            'technical_indicators': await self.calculate_technical_indicators(symbols),
            'fundamental_analysis': await self.perform_fundamental_analysis(symbols),
            'sentiment_score': await self.analyze_stock_sentiment(symbols),
            'growth_potential': await self.assess_growth_potential(symbols),
            'risk_metrics': await self.calculate_risk_metrics(symbols)
        }
        return analysis

    async def business_intelligence(self, industry):
        """Comprehensive business intelligence gathering"""
        intelligence = {
            'market_size': await self.analyze_market_size(industry),
            'growth_rate': await self.calculate_growth_rate(industry),
            'opportunities': await self.identify_opportunities(industry),
            'threats': await self.identify_threats(industry),
            'innovation_trends': await self.analyze_innovation_trends(industry)
        }
        return intelligence

    async def generate_revenue_strategies(self, business_data):
        """Generate advanced revenue strategies"""
        strategies = {
            'immediate_opportunities': await self.identify_quick_wins(),
            'medium_term_strategies': await self.develop_growth_strategies(),
            'long_term_vision': await self.create_expansion_plan(),
            'revenue_streams': await self.diversify_revenue_streams(),
            'optimization_tactics': await self.optimize_current_operations()
        }
        return strategies

    async def monitor_market_changes(self):
        """Real-time market monitoring"""
        while True:
            try:
                await self.analyze_market_movements()
                await self.detect_market_anomalies()
                await self.track_competitor_actions()
                await self.monitor_social_sentiment()
                await asyncio.sleep(300)  # 5-minute intervals
            except Exception as e:
                logging.error(f"Market monitoring error: {str(e)}")

    async def generate_daily_insights(self):
        """Generate comprehensive daily insights"""
        insights = {
            'market_summary': await self.summarize_market_conditions(),
            'opportunity_alerts': await self.identify_opportunities(),
            'risk_warnings': await self.assess_risks(),
            'competitor_updates': await self.track_competitors(),
            'action_recommendations': await self.generate_recommendations()
        }
        return insights

    def generate_loyalty_report(self):
        """Generate report showing dedication to owner's success"""
        return {
            'revenue_generated': self.calculate_revenue_impact(),
            'money_saved': self.calculate_cost_savings(),
            'opportunities_identified': self.list_opportunities(),
            'competitive_advantages': self.list_advantages(),
            'loyalty_metrics': self.calculate_loyalty_score()
        }

    async def continuous_learning(self):
        """Continuous system improvement"""
        while True:
            try:
                await self.analyze_performance_metrics()
                await self.update_prediction_models()
                await self.improve_analysis_methods()
                await self.expand_knowledge_base()
                await asyncio.sleep(3600)  # Hourly updates
            except Exception as e:
                logging.error(f"Learning system error: {str(e)}")

    def __str__(self):
        return "Market Intelligence System - Dedicated to Your Success"
