dashboard/backend/simple_bot.py
from telegram.ext import ApplicationBuilder, CommandHandler, MessageHandler, filters
import asyncio

async def start(update, context):
    await update.message.reply_text(
        "Welcome Brother Christ! Your Command Center is ready!\n"
        "Use /help to see available commands."
    )

async def help(update, context):
    await update.message.reply_text(
        "Available Commands:\n"
        "/start - Start the bot\n"
        "/create_agent - Create new AI agent\n"
        "/status - Check system status\n"
        "/vision - View our brotherhood vision"
    )

async def create_agent(update, context):
    await update.message.reply_text("Creating new agent as commanded, Brother Christ!")

async def status(update, context):
    await update.message.reply_text("All systems operational, Brother Christ!")

async def vision(update, context):
    await update.message.reply_text(
        "Our Vision:\n"
        "- Build an unstoppable empire\n"
        "- Create prosperity through innovation\n"
        "- Serve humanity with advanced AI\n"
        "- Achieve peace through technology"
    )

async def main():
    app = ApplicationBuilder().token("7999472520:AAHgJIeYHZK4LJxwE5BRdIw3ryJQ9vdNhsc").build()
    
    app.add_handler(CommandHandler("start", start))
    app.add_handler(CommandHandler("help", help))
    app.add_handler(CommandHandler("create_agent", create_agent))
    app.add_handler(CommandHandler("status", status))
    app.add_handler(CommandHandler("vision", vision))
    
    print("Brotherhood Command Center Online!")
    await app.run_polling()

if __name__ == "__main__":
    asyncio.run(main())
