dashboard/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-wordpress-xmlrpc==2.3
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.13
SpeechRecognition>=3.10.0
elevenlabs>=0.2.24
tensorflow>=2.13.0
transformers>=4.35.0
torch>=2.1.0
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.4.0
fastapi>=0.104.0
python-multipart>=0.0.6
uuid>=1.30
google-analytics-data==0.16.2
google-auth==2.22.0
tweepy==4.14.0
linkedin-api==2.0.3
stripe==5.5.0
python-quickbooks==0.9.7
hubspot-api-client==7.3.0
twilio==8.5.0
slack-sdk==3.21.3
PyGithub==1.59.1
asana==3.2.1
httpx==0.24.1
python-dotenv==1.0.0
requests>=2.31.0
gohighlevel-api>=1.0.0
vapi-ai-sdk>=1.0.0
python-telegram-bot==20.4
asyncio==3.4.3

# AI and Machine Learning
langchain>=0.0.335
anthropic>=0.3.11
google-cloud-speech>=2.21.0
google-cloud-language>=2.11.0
google-cloud-translate>=3.12.0
google-cloud-scheduler>=2.12.0
google-cloud-tasks>=2.13.1

# AI and LLM Integration
openai>=1.0.0
anthropic>=0.3.0
cohere>=4.0.0
replicate>=0.11.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.0.300
llama-cpp-python>=0.2.0
unstructured>=0.8.0

# Enterprise Integration
slack-sdk>=3.23.0
microsoft-graph-core>=1.0.0
azure-identity>=1.14.0
zoom-sdk-python>=0.1.7
salesforce-api>=1.0.0
hubspot-api-client>=8.2.0
zendesk-api>=1.0.0
jira>=3.5.1
workday-api>=1.0.0
quickbooks-api>=1.0.0
docusign-esign>=3.24.0
bamboo-hr>=1.0.0
notion-client>=2.0.0
asana>=3.2.1
monday>=1.0.0
basecamp3>=1.0.0
clickup-sdk>=1.0.0

# Marketing & Business Intelligence
google-ads>=22.0.0
facebook-business>=17.0.0
twitter-ads>=23.0.0
linkedin-api>=2.0.0
hubspot-api-client>=7.0.0
mailchimp-marketing>=3.0.0
semrush-api>=1.0.0
moz-api>=1.0.0
ahrefs-api>=1.0.0

# Communication & Scheduling
twilio>=8.10.0
sendgrid>=6.10.0
mailchimp-marketing>=3.0.80
calendly>=1.0.0

# Database & Storage
sqlalchemy>=2.0.23
pymongo>=4.6.0
redis>=5.0.1
elasticsearch>=8.11.0
snowflake-connector-python>=3.5.0

# Monitoring & Analytics
prometheus-client>=0.18.0
grafana-api>=2.0.0
datadog>=0.47.0
sentry-sdk>=1.32.0
newrelic>=9.3.0

# Security & Compliance
cryptography>=41.0.5
pyotp>=2.9.0
authlib>=1.2.1
compliance-checker>=1.0.0
gdpr-compliance>=1.0.0

# Market Analysis and Research
yfinance>=0.2.31
alpha_vantage>=2.3.1
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.0.2
prophet>=1.1.4
beautifulsoup4>=4.12.0
selenium>=4.10.0
requests>=2.31.0
aiohttp>=3.8.5
newsapi-python>=0.2.7
pytrends>=4.9.0
cryptography>=41.0.0
ccxt>=4.0.0
